refactor: change namespace 'background' to 'backgrounds'.
