+++ src/SS.ContentBundle/SContentBundleBuilder.Register.Backgrounds.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 
-using StardustSandbox.Core.Background;
+using StardustSandbox.Core.Backgrounds;
 using StardustSandbox.Core.Databases;
 using StardustSandbox.Core.Interfaces.General;
