+++ src/SS.Core/Backgrounds/Details/SCloud.cs
@@ -6,7 +6,7 @@ using StardustSandbox.Core.Interfaces.General;
 using StardustSandbox.Core.Mathematics;
 using StardustSandbox.Core.Objects;
 
-namespace StardustSandbox.Core.Background.Details
+namespace StardustSandbox.Core.Backgrounds.Details
 {
     public sealed class SCloud(ISGame gameInstance) : SGameObject(gameInstance), ISPoolableObject
     {
@@ -42,7 +42,7 @@ namespace StardustSandbox.Core.Background.Details
         {
             this.position = new Vector2(-this.texture.Width - SWorldConstants.GRID_SCALE, SRandomMath.Range(0, SWorldConstants.GRID_SCALE * 10));
             this.speed = SRandomMath.Range(10, 50);
-            this.opacity = ((float)SRandomMath.GetDouble() * 0.5f) + 0.5f;
+            this.opacity = (float)SRandomMath.GetDouble() * 0.5f + 0.5f;
         }
     }
 }
