+++ src/SS.Core/Backgrounds/SBackground.cs
@@ -7,7 +7,7 @@ using StardustSandbox.Core.Objects;
 
 using System.Collections.Generic;
 
-namespace StardustSandbox.Core.Background
+namespace StardustSandbox.Core.Backgrounds
 {
     public sealed class SBackground(ISGame gameInstance, string identifier, Texture2D texture) : SGameObject(gameInstance)
     {
