+++ src/SS.Core/Controllers/Background/SCloudController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
-using StardustSandbox.Core.Background.Details;
+using StardustSandbox.Core.Backgrounds.Details;
 using StardustSandbox.Core.Collections;
 using StardustSandbox.Core.Constants;
 using StardustSandbox.Core.Interfaces.General;
