+++ src/SS.Core/Managers/SBackgroundManager.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
-using StardustSandbox.Core.Background;
+using StardustSandbox.Core.Backgrounds;
 using StardustSandbox.Core.Controllers.Background;
 using StardustSandbox.Core.Interfaces.General;
 using StardustSandbox.Core.Objects;
